requirements.txt
requests==2.27.1
opencv-python==4.5.5.62
pyyaml==6.0